data_ingestion/requirements.txt
# Data Ingestion Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0

# Data handling
pandas==2.1.4
numpy==1.26.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Retry logic
tenacity==8.2.3

# Rate limiting
ratelimit==2.2.1

# Logging
coloredlogs==15.0.1

# JSON handling
ujson==5.9.0

# Optional: For advanced scraping
# scrapy==2.11.0
# selenium==4.16.0
